requirements.txt
django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pytest==8.4.2
python-dotenv==1.1.1
psycopg2-binary==2.9.11
redis==7.0.1